refactor(slack-app): move posthog-code activities into slack_app package
Splits the 17 @activity.defn functions out of posthog_code_slack_mention.py
and posthog_code_slack_mention_command.py into per-domain modules under
posthog/temporal/ai/slack_app/activities/ (billing, messaging, repo_selection,
rules, task_creation, thread, user_resolution). The two workflow files now
only hold workflow classes and import activities from the slack_app package.

derive_mention_workflow_id and POSTHOG_CODE_SLACK_MENTION_PICKER_GUIDANCE are
re-exported from the mention workflow module to preserve external call sites
in products/slack_app/backend/api.py and posthog/temporal/ai/eval_slack_repo_selection.py.

Worker registration in posthog/temporal/ai/__init__.py now unpacks
SLACK_APP_ACTIVITIES instead of listing each activity individually.
