+++ posthog/temporal/ai/__init__.py
@@ -9,29 +9,8 @@ from posthog.temporal.ai.posthog_code_slack_interactivity import (
     PostHogCodeSlackTerminateTaskWorkflow,
     process_posthog_code_terminate_task_activity,
 )
-from posthog.temporal.ai.posthog_code_slack_mention import (
-    PostHogCodeSlackMentionWorkflow,
-    block_posthog_code_task_if_no_personal_github_activity,
-    cascade_posthog_code_repository_activity,
-    classify_posthog_code_task_needs_repo_activity,
-    collect_posthog_code_thread_messages_activity,
-    create_posthog_code_routing_rule_activity,
-    create_posthog_code_task_for_repo_activity,
-    discover_posthog_code_repository_via_agent_activity,
-    enforce_posthog_code_billing_quota_activity,
-    forward_posthog_code_followup_activity,
-    handle_posthog_code_rules_command_activity,
-    post_posthog_code_internal_error_activity,
-    post_posthog_code_no_repos_activity,
-    post_posthog_code_picker_timeout_activity,
-    post_posthog_code_repo_picker_activity,
-    resolve_posthog_code_slack_user_activity,
-)
-from posthog.temporal.ai.posthog_code_slack_mention_command import (
-    PostHogCodeSlackMentionCommandWorkflow,
-    handle_posthog_code_slack_mention_command_activity,
-    resolve_posthog_code_slack_command_user_activity,
-)
+from posthog.temporal.ai.posthog_code_slack_mention import PostHogCodeSlackMentionWorkflow
+from posthog.temporal.ai.posthog_code_slack_mention_command import PostHogCodeSlackMentionCommandWorkflow
 from posthog.temporal.ai.research_agent import ResearchAgentWorkflow, process_research_agent_activity
 from posthog.temporal.ai.slack_app import SLACK_APP_ACTIVITIES
 from posthog.temporal.ai.slack_conversation import (
@@ -75,24 +54,7 @@ AI_ACTIVITIES = [
     process_research_agent_activity,
     summarize_llm_traces_activity,
     process_slack_conversation_activity,
-    enforce_posthog_code_billing_quota_activity,
-    resolve_posthog_code_slack_user_activity,
-    handle_posthog_code_rules_command_activity,
-    handle_posthog_code_slack_mention_command_activity,
-    resolve_posthog_code_slack_command_user_activity,
-    collect_posthog_code_thread_messages_activity,
-    create_posthog_code_routing_rule_activity,
-    cascade_posthog_code_repository_activity,
-    discover_posthog_code_repository_via_agent_activity,
-    classify_posthog_code_task_needs_repo_activity,
     *SLACK_APP_ACTIVITIES,
-    post_posthog_code_no_repos_activity,
-    post_posthog_code_repo_picker_activity,
-    block_posthog_code_task_if_no_personal_github_activity,
-    create_posthog_code_task_for_repo_activity,
-    forward_posthog_code_followup_activity,
-    post_posthog_code_picker_timeout_activity,
-    post_posthog_code_internal_error_activity,
     process_posthog_code_terminate_task_activity,
     investigate_anomaly_activity,
 ]
