+++ posthog/temporal/ai/posthog_code_slack_mention_command.py
@@ -1,15 +1,21 @@
 import json
 from datetime import timedelta
 
-from temporalio import activity, workflow
+from temporalio import workflow
 from temporalio.common import RetryPolicy
 
 from posthog.temporal.ai.slack_app import (
-    PostHogCodeSlackMentionCommandResult,
+    POSTHOG_CODE_SLACK_RULES_ADD_PICKER_GUIDANCE,
     PostHogCodeSlackMentionCommandWorkflowInputs,
+    PostHogCodeSlackMentionWorkflowInputs,
+    block_posthog_code_task_if_no_personal_github_activity,
+    create_posthog_code_routing_rule_activity,
+    handle_posthog_code_slack_mention_command_activity,
+    post_posthog_code_picker_timeout_activity,
+    post_posthog_code_repo_picker_activity,
+    resolve_posthog_code_slack_command_user_activity,
 )
 from posthog.temporal.common.base import PostHogWorkflow
-from posthog.temporal.common.utils import close_db_connections
 
 POSTHOG_CODE_SLACK_COMMAND_ACTIVITY_TIMEOUT_SECONDS = 60
 # Matches the mention workflow's picker wait window so the bot's behaviour is
@@ -46,15 +52,6 @@ class PostHogCodeSlackMentionCommandWorkflow(PostHogWorkflow):
 
     @workflow.run
     async def run(self, inputs: PostHogCodeSlackMentionCommandWorkflowInputs) -> None:
-        from posthog.temporal.ai.posthog_code_slack_mention import (
-            POSTHOG_CODE_SLACK_RULES_ADD_PICKER_GUIDANCE,
-            PostHogCodeSlackMentionWorkflowInputs,
-            block_posthog_code_task_if_no_personal_github_activity,
-            create_posthog_code_routing_rule_activity,
-            post_posthog_code_picker_timeout_activity,
-            post_posthog_code_repo_picker_activity,
-        )
-
         # New starts carry ``user_id`` from routing-time resolution and skip the
         # activity. Legacy histories started before the field existed deserialize
         # with ``user_id=None`` and replay through the activity so the recorded
@@ -158,129 +155,3 @@ class PostHogCodeSlackMentionCommandWorkflow(PostHogWorkflow):
             start_to_close_timeout=timedelta(seconds=POSTHOG_CODE_SLACK_COMMAND_ACTIVITY_TIMEOUT_SECONDS),
             retry_policy=RetryPolicy(maximum_attempts=3),
         )
-
-
-@activity.defn
-@close_db_connections
-def resolve_posthog_code_slack_command_user_activity(
-    inputs: PostHogCodeSlackMentionCommandWorkflowInputs,
-) -> int | None:
-    from posthog.models.integration import Integration, SlackIntegration
-
-    from products.slack_app.backend.api import resolve_posthog_user_from_event
-
-    event = inputs.event
-    channel = event.get("channel")
-    thread_ts = event.get("thread_ts") or event.get("ts")
-    slack_user_id = event.get("user")
-    if not channel or not thread_ts or not slack_user_id or not inputs.integration_ids:
-        return None
-
-    candidates = list(
-        Integration.objects.filter(
-            id__in=inputs.integration_ids,
-            kind="slack",
-            integration_id=inputs.slack_team_id,
-        ).select_related("team", "team__organization")
-    )
-    if not candidates:
-        return None
-
-    probe = candidates[0]
-    posthog_user = resolve_posthog_user_from_event(
-        slack_user_id=slack_user_id,
-        probe_integration=probe,
-        candidate_integrations=candidates,
-    )
-    if posthog_user is None:
-        SlackIntegration(probe).client.chat_postEphemeral(
-            channel=channel,
-            user=slack_user_id,
-            thread_ts=thread_ts,
-            text=(
-                "I couldn't find your PostHog account in any organization connected to this Slack "
-                "workspace. Ask an admin to invite you, then mention me again."
-            ),
-        )
-        return None
-    return posthog_user.id
-
-
-@activity.defn
-@close_db_connections
-def handle_posthog_code_slack_mention_command_activity(
-    inputs: PostHogCodeSlackMentionCommandWorkflowInputs,
-    user_id: int,
-) -> PostHogCodeSlackMentionCommandResult:
-    from posthog.models.integration import SlackIntegration
-
-    from products.slack_app.backend.api import _parse_rules_command
-    from products.slack_app.backend.services.commands import dispatch_rules_command, resolve_command_target
-
-    event = inputs.event
-    channel = event.get("channel")
-    thread_ts = event.get("thread_ts") or event.get("ts")
-    slack_user_id = event.get("user")
-    if not channel or not thread_ts or not slack_user_id:
-        return PostHogCodeSlackMentionCommandResult(status="done")
-
-    command = _parse_rules_command(event.get("text", ""))
-    if command is None:
-        return PostHogCodeSlackMentionCommandResult(status="done")
-
-    candidates, result = resolve_command_target(
-        slack_team_id=inputs.slack_team_id,
-        command=command,
-        slack_user_id=slack_user_id,
-        user_id=user_id,
-        channel=channel,
-        thread_ts=thread_ts,
-    )
-    if not candidates:
-        return PostHogCodeSlackMentionCommandResult(status="done")
-    if result.integration is None:
-        # Disambiguate "no access" (empty after access filtering) from
-        # "multiple projects available" so users get an actionable hint.
-        if not result.candidates:
-            text = (
-                "You don't have access to any PostHog project connected to this Slack workspace. "
-                "Ask an admin to grant you access, then try again."
-            )
-        else:
-            text = (
-                "This Slack workspace is connected to multiple PostHog projects. "
-                "Use `@PostHog project <id>` to set a default first, then re-run your command."
-            )
-        SlackIntegration(candidates[0]).client.chat_postEphemeral(
-            channel=channel,
-            user=slack_user_id,
-            thread_ts=thread_ts,
-            text=text,
-        )
-        return PostHogCodeSlackMentionCommandResult(status="done")
-
-    target = result.integration
-
-    # ``rules add`` without an inline repo needs the interactive picker, which
-    # only a workflow can drive (it owns the signal). Hand control back so the
-    # workflow can post the picker against the resolved target and wait for
-    # the user's selection.
-    if command.action == "add" and not command.repository:
-        return PostHogCodeSlackMentionCommandResult(
-            status="needs_picker",
-            pending_rule_text=command.rule_text,
-            target_integration_id=target.id,
-        )
-
-    dispatch_rules_command(
-        command,
-        SlackIntegration(target),
-        target,
-        channel=channel,
-        thread_ts=thread_ts,
-        slack_user_id=slack_user_id,
-        slack_workspace_id=inputs.slack_team_id,
-        user_id=user_id,
-        workspace_candidates=candidates,
-    )
-    return PostHogCodeSlackMentionCommandResult(status="done")
