+++ posthog/temporal/ai/slack_app/activities/__init__.py
@@ -1,11 +1,61 @@
+from posthog.temporal.ai.slack_app.activities.billing import enforce_posthog_code_billing_quota_activity
 from posthog.temporal.ai.slack_app.activities.classifiers import (
     CLASSIFIER_THREAD_HISTORY_MESSAGES,
     classify_message_is_agent_directed,
     classify_untagged_followup_activity,
 )
+from posthog.temporal.ai.slack_app.activities.messaging import (
+    POSTHOG_CODE_SLACK_MENTION_PICKER_GUIDANCE,
+    POSTHOG_CODE_SLACK_RULES_ADD_PICKER_GUIDANCE,
+    block_posthog_code_task_if_no_personal_github_activity,
+    post_posthog_code_internal_error_activity,
+    post_posthog_code_no_repos_activity,
+    post_posthog_code_picker_timeout_activity,
+    post_posthog_code_repo_picker_activity,
+)
+from posthog.temporal.ai.slack_app.activities.repo_selection import (
+    cascade_posthog_code_repository_activity,
+    classify_posthog_code_task_needs_repo_activity,
+    discover_posthog_code_repository_via_agent_activity,
+)
+from posthog.temporal.ai.slack_app.activities.rules import (
+    create_posthog_code_routing_rule_activity,
+    handle_posthog_code_rules_command_activity,
+    handle_posthog_code_slack_mention_command_activity,
+)
+from posthog.temporal.ai.slack_app.activities.task_creation import (
+    create_posthog_code_task_for_repo_activity,
+    derive_mention_workflow_id,
+    forward_posthog_code_followup_activity,
+)
+from posthog.temporal.ai.slack_app.activities.thread import collect_posthog_code_thread_messages_activity
+from posthog.temporal.ai.slack_app.activities.user_resolution import (
+    resolve_posthog_code_slack_command_user_activity,
+    resolve_posthog_code_slack_user_activity,
+)
 
 __all__ = [
     "CLASSIFIER_THREAD_HISTORY_MESSAGES",
+    "POSTHOG_CODE_SLACK_MENTION_PICKER_GUIDANCE",
+    "POSTHOG_CODE_SLACK_RULES_ADD_PICKER_GUIDANCE",
+    "block_posthog_code_task_if_no_personal_github_activity",
+    "cascade_posthog_code_repository_activity",
     "classify_message_is_agent_directed",
+    "classify_posthog_code_task_needs_repo_activity",
     "classify_untagged_followup_activity",
+    "collect_posthog_code_thread_messages_activity",
+    "create_posthog_code_routing_rule_activity",
+    "create_posthog_code_task_for_repo_activity",
+    "derive_mention_workflow_id",
+    "discover_posthog_code_repository_via_agent_activity",
+    "enforce_posthog_code_billing_quota_activity",
+    "forward_posthog_code_followup_activity",
+    "handle_posthog_code_rules_command_activity",
+    "handle_posthog_code_slack_mention_command_activity",
+    "post_posthog_code_internal_error_activity",
+    "post_posthog_code_no_repos_activity",
+    "post_posthog_code_picker_timeout_activity",
+    "post_posthog_code_repo_picker_activity",
+    "resolve_posthog_code_slack_command_user_activity",
+    "resolve_posthog_code_slack_user_activity",
 ]
